+++ src/app/types/general.types.ts
@@ -24,20 +24,52 @@ export interface DialogData {
   source: string;
 }
 
+export enum IdentifiersType {
+  DOI = 'doi',
+  ISBN = 'isbn',
+}
+
 export class Filters {
   title: string;
   author: string;
   feed: string;
-  constructor(title: string = '', author: string = '', feed: string = '') {
+  from: string = '';
+  to: string = '';
+  constructor(
+    title: string = '',
+    author: string = '',
+    feed: string = '',
+    from: string[] = [],
+    to: string[] = []
+  ) {
     this.title = title;
     this.author = author;
     this.feed = feed;
+    for (let i = 0; i < from.length; i++) {
+      if (from[i]) {
+        if (i === 0) {
+          this.from += from[i];
+        } else {
+          this.from += `-${from[i]}`;
+        }
+      }
+    }
+    for (let i = 0; i < to.length; i++) {
+      if (to[i]) {
+        if (i === 0) {
+          this.to += to[i];
+        } else {
+          this.to += `-${to[i]}`;
+        }
+      }
+    }
   }
   getFilters() {
-    return `title=${this.title}&feed=${this.feed}&author=${this.author}`;
+    return `title=${this.title}&feed=${this.feed}&author=${this.author}&from=${this.from}&to=${this.to}`;
   }
   isActive() {
-    if (this.title || this.author || this.feed) return true;
+    if (this.title || this.author || this.feed || this.from || this.to)
+      return true;
     return false;
   }
 }