+++ src/app/types/my-shelf.types.ts
@@ -1,27 +1,27 @@
 import { Metadata } from './general.types';
 import { EntryDetail } from './entry.types';
 
-export interface FavoriteList {
-  items: Favorite[];
+export interface MyShelfList {
+  items: MyShelf[];
   metadata: Metadata;
 }
 
-export interface Favorite {
+export interface MyShelf {
   id: string;
   entry: EntryDetail;
   created_at: string;
   updated_at: string;
 }
 
-export interface FavoriteNew {
+export interface MyShelfNew {
   entry_id: string;
 }
 
-export interface FavoritePostResponse {
-  response: Favorite;
+export interface MyShelfPostResponse {
+  response: MyShelf;
 }
 
-export interface FavoriteQuery {
+export interface MyShelfQuery {
   page: number;
   limit: number;
   title: string;