+++ src/environments/environment.prod.ts
@@ -3,5 +3,4 @@ export const environment = {
   baseUrl: `https://evilflowers.fiit.stuba.sk`,
   assetsUrl: `https://elvira.fiit.stuba.sk/assets`,
   catalog_id: `b349b5a1-4d50-4bed-bf8d-c56d6c7b1280`,
-  httpRequestInterval: 300000, // 300s
 };